♻️  zb: Split object_server::DispatchNotifier into separate mod
